+++ changelog.md
@@ -1,3 +1,12 @@
+# v1.1.3
+- Changed Android32 to NOT use TextArea (geode needs to fix gd::string!) basically setting a description is now fixed with a few caviats
+- Made downloading pending objects not send a request.
+- Updated Editor Tab API dependency ver
+- Added debug logs because why not
+- Made scroll layer cancelling touches less sensitive
+- Fixed issue with downloading/favoriting objects sometimes causing a crash when you exit too quickly
+# v1.1.2
+- Added some safe guards
 # v1.1.1
 - Made typing in pages work
 - Made pressing escape while in an object not close the menu, but instead "go back"